% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlent.R
\docType{data}
\name{ents}
\alias{ents}
\title{HTML entities}
\format{A list of 1448 HTML entities}
\source{
\url{https://dev.w3.org/html5/html-author/charref}
}
\usage{
ents
}
\description{
HTML entities
}
\keyword{datasets}
